Separate simulation components and unify UI layout
- Created BaseSelectionSliderLayout.tsx in parts for shared slider UI
- Restored OptionSliderControl logic and refactored it to use BaseSelectionSliderLayout
- Created SimulationSliderControl with its own logic, also using BaseSelectionSliderLayout
- Updated SimulationDialog to use the domain-specific SimulationSliderControl
- Deleted global useSelectionSlider hook to keep logic local to domain components
- Fixed Biome linting and formatting issues
- Verified all tests and coverage report generation
